requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
python-jose
passlib
bcrypt